+++ src/index.html
@@ -1,9 +1,9 @@
 <!DOCTYPE html>
-<html lang="ru">
+<html lang="en">
   <head>
     <meta charset="UTF-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>Page title</title>
+    <title>Ice cream</title>
 
     <!-- font-family -->
     <link rel="preconnect" href="https://fonts.googleapis.com" />
