+++ src/js/gallery.js
@@ -1,0 +1,4 @@
+$('.single-item').slick({
+    autoplay: true,
+    arrows: false,
+});
